ci: only run on push to main branch
Currently, "pace main unit tests" run on push to any branch. In case of
`develop`, this is redundant with the merge queue checks. In case of any
branch that isn't protected (e.g. isn't the main branch), running ci
isn't necessary. This is consistent with the linting workflow and other
workflow configurations (e.g. PyFV3 and PySHiELD).