Use eta files shipped with NDSL (#149)
Some tests and examples in pace depend on eta files for 79 and 91
layers. These eta files are now shipped with NDSL starting with

https://github.com/NOAA-GFDL/NDSL/pull/171

Since pace depends on NDSL, there's no need for a script to generate
these eta files: They are already present in the NDSL/ directory that is
pulled in as a submodule. Tests and examples thus don't have to generate
eta files anymore before they can run. This simplifies workflows (e.g.
the GHA one), but also on-boarding of new users.

The PR includes a buch of typo fixes and markdown improvements.

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>